+++ test/lightning/kafka_triggers/event_listener_test.exs
@@ -5,6 +5,7 @@ defmodule Lightning.KafkaTriggers.EventListenerTest do
   alias Lightning.KafkaTriggers.EventListener
   alias Lightning.KafkaTriggers.PipelineSupervisor
   alias Lightning.Workflows.Triggers.Events
+  alias Lightning.Workflows.Triggers.Events.KafkaTriggerNotificationSent
   alias Lightning.Workflows.Triggers.Events.KafkaTriggerUpdated
 
   import Mox
@@ -40,6 +41,18 @@ defmodule Lightning.KafkaTriggers.EventListenerTest do
     Events.kafka_trigger_updated(trigger_id)
 
     assert_receive %KafkaTriggerUpdated{trigger_id: ^trigger_id}
+  end
+
+  test "init/1 subscribes to the Kafka notifications topic", %{
+    trigger: trigger
+  } do
+    trigger_id = trigger.id
+
+    assert EventListener.init([]) == {:ok, %{}}
+
+    Events.kafka_trigger_notification_sent(trigger_id, ~U[2021-01-01 12:00:00Z])
+
+    assert_received %KafkaTriggerNotificationSent{trigger_id: ^trigger_id}
   end
 
   test "handle_info/1 updates the trigger's pipeline", %{
@@ -99,5 +112,59 @@ defmodule Lightning.KafkaTriggers.EventListenerTest do
 
       assert_not_called(KafkaTriggers.update_pipeline(:_, :_))
     end
+  end
+
+  test "handle_info/1 sets the notification sent at time if not set", %{
+    state: state,
+    trigger: %{id: trigger_id}
+  } do
+    sent_at = DateTime.utc_now()
+
+    assert {:noreply, ^state} =
+             EventListener.handle_info(
+               %KafkaTriggerNotificationSent{
+                 trigger_id: trigger_id,
+                 sent_at: sent_at
+               },
+               state
+             )
+
+    persisted_sent_at =
+      :persistent_term.get(
+        {:kafka_trigger_failure_notification_sent_at, trigger_id},
+        nil
+      )
+
+    assert persisted_sent_at == sent_at
+  end
+
+  test "handle_info/1 updates the notification sent at time if set", %{
+    state: state,
+    trigger: %{id: trigger_id}
+  } do
+    sent_at = DateTime.utc_now()
+    last_sent_at = DateTime.add(sent_at, -30, :second)
+
+    :persistent_term.put(
+      {:kafka_trigger_failure_notification_sent_at, trigger_id},
+      last_sent_at
+    )
+
+    assert {:noreply, ^state} =
+             EventListener.handle_info(
+               %KafkaTriggerNotificationSent{
+                 trigger_id: trigger_id,
+                 sent_at: sent_at
+               },
+               state
+             )
+
+    persisted_sent_at =
+      :persistent_term.get(
+        {:kafka_trigger_failure_notification_sent_at, trigger_id},
+        nil
+      )
+
+    assert persisted_sent_at == sent_at
   end
 end