+++ test/lightning/kafka_triggers/pipeline_test.exs
@@ -14,6 +14,8 @@ defmodule Lightning.KafkaTriggers.PipelineTest do
   alias Lightning.KafkaTriggers.TriggerKafkaMessageRecord
   alias Lightning.KafkaTriggers.Pipeline
   alias Lightning.Repo
+  alias Lightning.Workflows.Triggers.Events
+  alias Lightning.Workflows.Triggers.Events.KafkaTriggerNotificationSent
   alias Lightning.WorkOrder
 
   describe ".start_link/1" do
@@ -392,8 +394,15 @@ defmodule Lightning.KafkaTriggers.PipelineTest do
         build_broadway_message(offset: 2) |> Broadway.Message.failed(:who_knows)
       ]
 
+      trigger =
+        insert(
+          :trigger,
+          type: :kafka,
+          kafka_configuration: build(:triggers_kafka_configuration)
+        )
+
       %{
-        context: %{trigger_id: "my_trigger_id"},
+        context: %{trigger_id: trigger.id |> String.to_atom()},
         messages: messages
       }
     end
@@ -440,6 +449,29 @@ defmodule Lightning.KafkaTriggers.PipelineTest do
       end
     end
 
+    test "notifies the project admin of a persistence failure", %{
+      context: context,
+      messages: [message_1, message_2]
+    } do
+      trigger_id = context.trigger_id |> Atom.to_string()
+
+      persistence_failure_message =
+        build_broadway_message(offset: 3)
+        |> Broadway.Message.failed(:persistence)
+
+      messages = [
+        message_1,
+        persistence_failure_message,
+        message_2
+      ]
+
+      Events.subscribe_to_kafka_trigger_updated()
+
+      Pipeline.handle_failed(messages, context)
+
+      assert_receive %KafkaTriggerNotificationSent{trigger_id: ^trigger_id}
+    end
+
     defp expected_duplicate_log_message(message, context) do
       "Kafka Pipeline Duplicate Message:" <>
         " Trigger_id `#{context.trigger_id}`" <>
@@ -471,6 +503,66 @@ defmodule Lightning.KafkaTriggers.PipelineTest do
         trigger_id: context.trigger_id,
         type: type
       }
+    end
+  end
+
+  describe "maybe_notify_users/2" do
+    setup do
+      Events.subscribe_to_kafka_trigger_updated()
+
+      trigger =
+        insert(
+          :trigger,
+          type: :kafka,
+          kafka_configuration: build(:triggers_kafka_configuration)
+        )
+
+      %{
+        trigger_id: trigger.id
+      }
+    end
+
+    test "does not notify the user if there are no persistence failures", %{
+      trigger_id: trigger_id
+    } do
+      messages = [
+        build_broadway_message() |> Broadway.Message.failed(:other),
+        build_broadway_message() |> Broadway.Message.failed(:other)
+      ]
+
+      Pipeline.maybe_notify_users(messages, trigger_id)
+
+      refute_receive %KafkaTriggerNotificationSent{}
+    end
+
+    test "notifies users if there is a message with a persistence failure", %{
+      trigger_id: trigger_id
+    } do
+      messages = [
+        build_broadway_message() |> Broadway.Message.failed(:other),
+        build_broadway_message() |> Broadway.Message.failed(:persistence),
+        build_broadway_message() |> Broadway.Message.failed(:other)
+      ]
+
+      Pipeline.maybe_notify_users(messages, trigger_id)
+
+      assert_receive %KafkaTriggerNotificationSent{trigger_id: ^trigger_id}
+    end
+
+    test "notifies users if there are multiple persistence failures", %{
+      trigger_id: trigger_id
+    } do
+      messages = [
+        build_broadway_message() |> Broadway.Message.failed(:other),
+        build_broadway_message() |> Broadway.Message.failed(:persistence),
+        build_broadway_message() |> Broadway.Message.failed(:other),
+        build_broadway_message() |> Broadway.Message.failed(:persistence),
+        build_broadway_message() |> Broadway.Message.failed(:other)
+      ]
+
+      Pipeline.maybe_notify_users(messages, trigger_id)
+
+      assert_receive %KafkaTriggerNotificationSent{trigger_id: ^trigger_id}
     end
   end